+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed MethodInputParametersMustBeFinal to correctly handle enum types
+
 ## [1.0.0] - 2024-03-25
 ### Added
 - Extracting Besu custom error-prone checks to its own repository